refactor: extract github comment section helpers
